+++ src/Track/track.service.ts
@@ -1,17 +1,6 @@
 import { v4 as uuid } from 'uuid';
-import { Injectable, forwardRef, Inject } from '@nestjs/common';
-import { DBEntity } from '../utils/DB/DBEntyty';
-import { OmitType, PartialType } from '@nestjs/mapped-types';
-import {
-  IsNotEmpty,
-  IsUUID,
-  IsString,
-  ValidateIf,
-  IsInt,
-} from 'class-validator';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { ErrorsCode } from '../utils/common types/enum';
-import { AlbumService } from '../Album/album.service';
-import { ArtistService } from '../Artist/artist.service';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { TrackEntity } from './entities/track.entity';
@@ -24,18 +13,14 @@ export class TrackService {
     @InjectRepository(TrackEntity)
     private trackRepository: Repository<TrackEntity>,
   ) {}
+
   async create(trackDto: CreateTrackDTO) {
-    const created: TrackEntity = {
-      ...trackDto,
-      id: uuid(),
-    };
-    const createdTrack = this.trackRepository.create(created);
+    const createdTrack = this.trackRepository.create(trackDto);
     return await this.trackRepository.save(createdTrack);
   }
 
   async findAll() {
     const tracks = await this.trackRepository.find();
-
     return tracks;
   }
 
@@ -44,8 +29,8 @@ export class TrackService {
       where: { id: trackId },
     });
 
-    if (track) return track;
-    throw new Error(ErrorsCode[404]);
+    if (!track) throw new NotFoundException(ErrorsCode[404]);
+    return track;
   }
 
   async change(trackId: string, chandeDTO: ChangeTrackDTO) {
@@ -54,7 +39,7 @@ export class TrackService {
     });
 
     if (!trackForUpdate) {
-      throw new Error(ErrorsCode[404]);
+      throw new NotFoundException(ErrorsCode[404]);
     }
 
     const changedEntity = { ...trackForUpdate, ...chandeDTO };
@@ -63,10 +48,14 @@ export class TrackService {
   }
 
   async delete(trackId: string) {
-    const deletedTrack = await this.trackRepository.delete(trackId);
+    const trackForUpdate = await this.trackRepository.findOne({
+      where: { id: trackId },
+    });
 
-    if (deletedTrack.affected === 0) {
-      throw new Error(ErrorsCode[404]);
+    if (!trackForUpdate) {
+      throw new NotFoundException(ErrorsCode[404]);
     }
+
+    return await this.trackRepository.delete(trackId);
   }
 }
